Crawler/requirements.txt
apscheduler==3.11.2
pymongo==4.17.0
python-dotenv==1.2.2
selenium==4.43.0
